refactor(invite): move the invite procedures into their own router (#2877)
The 6 invite procedures sat in the middle of organization.ts, a 1535-line file
whose other 10 procedures are organization CRUD. They now live in invite.ts,
and organization.ts drops to 787 lines.

The procedures are merged back under the organization key in root.ts with
mergeRouters, so every one of them keeps its existing organization.* path. That
is what satisfies the ticket's "no API contract changes": mounting the new
router at its own invite key would have moved 6 paths and forced ~52 caller
updates across the frontend. If a real invite.* namespace is wanted, that is a
follow-up whose cost is those callers, not this move.

The 714 moved lines are byte-for-byte unchanged, so the diff is a move rather
than a rewrite. Three edits around them:

- teamRoleInputSchema was a file-local const used by procedures on both sides of
  the split, so it moves to teamRole.schemas.ts and both routers import it. Its
  two building blocks stay private to that module, since only the union is used
  at call sites.
- organization.ts loses the 8 imports that only the invite procedures used.
- mergeRouters is exported from trpc.ts, with a note on when to use it over
  mounting a router in root.ts.

One test changes, and it is unavoidable under any extraction: the acceptInvite
unit test builds a caller from organizationRouter directly, so moving the
procedure out of that router breaks the harness regardless of the merged path.
It now builds its caller from inviteRouter. The file keeps its name and its
"organization.acceptInvite" describe, because the path it exercises is still
organization.acceptInvite.
